Fix an issue where C++ include header is in lowercase (#221)
The basename is used to generate the real files, including header files. When generating the include statement, the file name should be exactly the same as the physical file name, otherwise, it cannot be found on Linux, which is a case-sensitive OS.

For example, if the file name is 'BaseName_interface.hpp', the include statement should be '#include "BaseName_interface.hpp"' rather than '#include "basename_interface.hpp"'.

Co-authored-by: Oliver Wu <oliver.wu@autodesk.com>